templates/packer: don't subscribe fedora executor

+++ templates/packer/ansible/roles/common/files/worker-initialization-scripts/worker_executor.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 set -euo pipefail
 
-/usr/local/bin/aws secretsmanager get-secret-value \
-  --secret-id executor-subscription-manager-command | jq -r ".SecretString" > /tmp/subscription_manager_command.json
-jq -r ".subscription_manager_command" /tmp/subscription_manager_command.json | bash
-rm -f /tmp/subscription_manager_command.json
+# Don't subscribe on fedora
+source /etc/os-release
+if [ "$ID" != fedora ]; then
+  /usr/local/bin/aws secretsmanager get-secret-value \
+    --secret-id executor-subscription-manager-command | jq -r ".SecretString" > /tmp/subscription_manager_command.json
+  jq -r ".subscription_manager_command" /tmp/subscription_manager_command.json | bash
+  rm -f /tmp/subscription_manager_command.json
+fi
 
 echo "Starting osbuild-jobsite-builder."
 /usr/libexec/osbuild-composer/osbuild-jobsite-builder
